Move io functionality to it own module
